Predicated reflection and JNI config for NI
Introducing predicates for reflection and JNI configuration. The predicate
can be only a type name. For example,
"predicate": {
  "whenTypeReachable": "org.graalvm.ReachableClass"
 },

will apply configuration only if "type org.graalvm.ReachableClass" is reachable.
The default predicate for every reflection entry is "java.lang.Object" and as
such it will not be printed.

The Native Image agent does not emit predicates, but will fuse predicates
accordingly. Instead all config of one type, the fusion will happen by
the predicate and the type.